Simplify namespace only shards and user entries
